sim refactor & cleanup
